Add Cloudflare Pages Function for iframe preflight API
PR previews deploy static assets to Cloudflare Pages, so Vercel-only
api/check-iframe was never served. Mirror the endpoint with a Pages
Function and share logic via lib/iframe-preflight.js.
